mediapipe_face/requirements.txt
opencv-python
mediapipe
pymongo
python-dotenv
numpy
requests
flask
flask-cors
scikit-learn
